refactor(routes): move outbound pushes out of caseLifecycle
caseLifecycle.ts was frozen at 1204 lines by the size ledger, and it fought two
of the security fixes earlier in this series: the atomic case-claim had to pay
for five new lines by deleting comments elsewhere, and the export snapshot had
to be squeezed onto one line. A ledger that makes a security fix arrive with
less documentation is working against the thing it exists to protect.

The largest cohesive block in it was the outbound pushes — DFIR-IRIS,
Timesketch, MISP, Notion, ClickUp, Jira, ServiceNow, plus the status and
reconnect endpoints they depend on. 417 lines, and the least related to
everything around them: this module is about a case's OWN lifecycle (create,
archive, restore, delete, export) while every push route is about a third-party
system's API, its credentials and its failure modes. A Jira change and a
case-deletion change no longer share a file.

MOVED VERBATIM. Route bodies, status codes and log lines are unchanged; only the
registration point moved, following the registerSeedDemoRoutes /
registerEncryptedImportRoutes pattern already used here. logLine is re-declared
in the new module for the same reason caseLifecycle declares it — so 30 moved
call sites stay byte-identical instead of being rewritten to serverLogger.info.

caseLifecycle.ts: 1204 -> 791, which is under the 800-line threshold, so it
leaves the ledger ENTIRELY rather than being re-recorded at a smaller number.
The ledger goes 13 -> 12 files.

The stale bullet in caseLifecycle's own docblock claiming to own the push routes
is corrected rather than left to mislead the next reader.

Verified: typecheck, eslint, prettier clean; server + routes suites 1056 passed;
boundary, import-cycle and a11y ratchets unmoved.
